videodev: fix a NULL pointer dereference in v4l2_device_release()
The change in 8280b66 does not cover the case when v4l2_dev is already
NULL, fix that.

With a Kinect sensor, seen as an USB camera using GSPCA in this context,
a NULL pointer dereference BUG can be triggered by just unplugging the
device after the camera driver has been loaded.

Signed-off-by: Antonio Ospite <ospite@studenti.unina.it>
Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>
Change-Id: I29033879be721b3250228a1f6ab5897267bed330
Reviewed-on: http://gerrit.pcs.mot.com/407935
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Ravikumar Vembu <a18394@motorola.com>