IKKRNBSP-932 Updates to the LM75 temp driver
Need to make updates to utilize the temp sensor's fault queue for a more
robust data set. Additionally we are modifying the irq handler to
invalidate a flag with each interrupt, forcing the algorithm to send a
read each time; this corrects an issue where the interrupts are coming
very quickly and the temp data reported back can be out of sync
with reality.
Porting IKHSS7-4122 to main-dev-ics.

Change-Id: I98fc6fa4fac9f19b69ae8395cafe0fcbefd7428b
Reviewed-on: http://gerrit.pcs.mot.com/361915
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>