msm: pm-8x60: Warmboot support for 9615
During warmboot, the boot remap register in the A5 core allows for a 64K
aligned warmboot handler to be remapped to the boot address. Use the
reserve callback in the machine_descriptor to allocate a memory block
that is aligned to 64K using memblock_alloc API.  The boot remapper
address is programmed so as to remap the allocated memory to boot
address. The power management code programs into this memory to jump to
the warmboot resume path during power collapse.

Change-Id: I2cf9e5ca346ac0632c9e21152dc25f5541166077
Signed-off-by: Maheshkumar Sivasubramanian <msivasub@codeaurora.org>
[jeff.carlyle@motorola.com: adapted from c6c5503871a2f8b86c0e3a203f7b389bdd6c1307]
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>