power: pm8921-bms: force max_voltage at end of charge
When an end of charge with a full battery happens, the bms driver
fakes max ocv and zero cc. To force max ocv the driver currently
assumes that the first row last column will have the highest profiled
voltage. This may not be true always.

Pass the maximum voltage from platform data and use that to force
max ocv.

Change-Id: I29c71a20648fb2f9066f2f82e14a6080692c63fd
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
[jeff.carlyle@motorola.com: adapted from a93ede8e3f073b4b9f8999afa1ff782cca643547]
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>