IKQCOMSD1-349: turn off DEBUG_USER option
Having this on is neat sometimes, but if some userspace thing
keeps crashing and restarting, then all that kernel does is
dumping those stacks endlessly, becoming unresponsive.

It also interferes with certain benchmarks that inject faults on
purpose, such as lmbench. This should not be on in general, as
we have tombstones for userland crashes.

Change-Id: I23db0a0d1b13bbe2ce2de3813677e89a87df1247
Signed-off-by: Igor Kovalenko <cik009@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/309334
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>