IKQCOMSD1-493 Temporarily Disable HDCP
An HDCP issue exists such that no non-black frames are rendered
over HDMI.  To avoid this issue, HDCP will be disabled until a
proper fix for the HDCP issue can be found.

A second HDCP related issue is worked around with this change.
A 20ms delay is added to avoid a lock up that can occur when
connecting to some HDMI monitors.  This is again a short term
work-around until a proper fix can be found

Change-Id: I304a4aa69283c1aa5f6329252175e206f55b309d
Reviewed-on: http://gerrit.pcs.mot.com/314288
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>