msm: spm-8960: Initialize VCTL to 1.3V at init
The SAW_VCTL register controls the voltage of the krait rails.
Initialize the krait rails to a 1.3V as that is the new maximum
voltage.

Otherwise, we may under volt the CPU if its currently running at
a higher frequency that needs a higher voltage and we
inadvertently lower the voltage before the cpu clock driver can
come online and increase the voltage to what is required. Right
now the bootloaders are booting at 1.3V on S5 and S6 but the spm
code is bringing it back down to 1.05V.

Change-Id: Ib08c747608d1fe8843609e74decb8aa73e0eaa3b
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
[jeff.carlyle@motorola.com: adapted from 9ac587a55b8fbb8ab80874c4b559f5782b4fbd4a]
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>