IKVANQUISH-256: Leave otg state as undefined until emu detection
The otg state needs to be left in the undefined state until the initial
emu detection is completed. Setting the state to idle within the otg
probe function allows the USB controller to be powered down. The device
controller driver's probe fails as a result which eventually leads to
a kernel panic.

Change-Id: Ia4d4e314c584859f86efbd7ef3a2b8e1a89695f0
Signed-off-by: Neil Patel <NPATEL@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/320099
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>