msm: board-8960: Wrap PMIC 8921 GPIO 42 in a gpio-regulator device
Wrap PMIC 8921 GPIO 42 in a gpio-regulator device so that the USB
OTG driver can be simplified.  GPIO 42 represents the OTG_CNTL
signal on 8960 devices.  This drives a FET which switches the USB
OTG MVS output of the PMIC 8921.  By controlling GPIO 42 with a
gpio-regulator, it is possible to encapsulate all USB OTG 5V
output dependencies within a single regulator_enable call. The
regulator framework is able to automatically set all GPIOs and
PMIC regulators appropriately thanks to the supply chain that
is defined for the regulators.

Change-Id: I8cd26b56a3a2871e784227b5d49f5d1089b19784
Signed-off-by: David Collins <collinsd@codeaurora.org>
[jeff.carlyle@motorola.com: adapted from 35a78f54e8b57dcd96b815b448e1a7696a4ae55d]
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>