IKQCOMSD1-1047: Set phy capabilites for ACCY_CONTROL mode
Set the phy caps so that we allow phy retention, phy power collapse, and
disabling of the OTG comparators. Since the EMU Detection driver is relying on
the PMIC and GPIOs to trigger accessory detection we can enable these flags for
increased power savings. Also the EMU Detection driver needs to verify that the
OTG driver resume completes before it attempts to read USB controller registers

Change-Id: I8dd4c767a76fcffc6d0111c30800509185c0f468
Signed-off-by: Neil Patel <NPATEL@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/335972
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>