IKQCOMSD1-1149: Fix msm_otg_set_host failure at boot
This change reverts most of commit 8f3c8bb5cab2f7b6ac3fa744d5dca9dc74920240.
The change caused a failure in msm_otg_set_host during the OTG driver init
which resulted in USB Host not working as well as enumeration to fail on
HW that does not have working EMU detection circuitry. With this minimal
change we are still able to see smart dock undetection happen which was
the main goal of the original change.

Change-Id: I137ed9d232d07a75b02ec81f9d779a768e25ed94
Signed-off-by: Neil Patel <NPATEL@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/336752
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>