IKHSS7-5509:Prop:IKHSS6-3415:Fix Caps lock indication issue with design fix
The previous design using the HID Core driver design element created
multiple virtual devices for each report id type and one each for input
and output for a single report id. So, the caps lock key was configured
under one virtual device and the caps lock LED under another. So when
the caps lock key was pressed on the first virtual device, the LED
indication was never sent back since that virtual device was not
configured for LEDs.

So the fix now would be to not use the base HID core driver design
to create multiple virtual devices but create just one using the
HID core driver and then create another in the hid-motorola.c driver
to be used for the touch inputs.

Also, revert the changes done to temporarily fix this issue under
IKTABLETMAIN-11091

Change-Id: I71dd70ce7b6c501421b598306ed76e3cb171bb3e
Signed-off-by: Shilpa B S <qfgk43@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/343243
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: LieJun Tao <L.J.Tao@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>