IKQCOMSD1-899 Avoid resetting the phy while the OTG driver is still host mode
         Upon disconnection of the SMART DOCK, the EMU detection driver is resetting
the USB PHY before instructing it to come out of host mode. This is an unexpected
condition for the OTG driver. The emu detection driver should prevent the reset from
happening in cases where the PHY in in USB host mode.

          This also fixes the headset detection issue where the detection driver needs
to know the interrupt occured, because the headset detection is evaluated only on the
occurence of ID interrupt

Change-Id: I1511880eba1554e508196f1674407871a5b469f1
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/338405
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Neil Patel <NPATEL@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Keyur Khambholja <Keyur.Khambholja@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>