IKQCOMSD1-1119: Fix Eprom Read Issue on Older HW
The is a conflict on older HW designs where the L7
Regulator is shared across multiple things.  1-wire
needs the level to be 2.7 V or higher and the Display
needs it to 1.85 V.  To fix this conflict and ensure
1-wire validity the level shifter is being moved to
using an internal pullup of 1K ohm to L17 this mimics
the change that was done on newer HW revs.

Change-Id: I7bd97bf287a6c7ff337359275f31b96f86b60537
Signed-off-by: w65789 <w65789@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/339896
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Keyur Khambholja <Keyur.Khambholja@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>