IKQCOMSD1-1029: Force Shutdown at 3.2 V with Cable attached
The testing for this part of the feature did not work
before but appears that it is functioning now.  This
change forces the reporting of ac and usb to 0 when
the voltage comparator dips below 3.2V.  This forces
Android to properly power down even while attached
to the accessory.  The phone should then enter Dead
Battery Recovery before powering into either Charge
Only Mode or Normal Android.

Change-Id: I9d25c1ab817c6ea546c13e953990dac7b0d04270
Signed-off-by: w65789 <w65789@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/340343
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>