IKHSS7-5668: fix ov8820 camera probe
The match_id function is called during driver probe.  The return
value of the function determines probe success/failure.  A
positive return value is not an error, but is nevertheless
interpreted as one in dd.c, resulting in a bogus probe failure
message in the log.  Note: the probe actually succeeds; this
is only to prevent the bogus log messge.  Additionally, add a
warm/fuzzy success confirmation log to be consistent with other
sensor drivers.

Change-Id: I6987a9795b1240eb4e71baf55c9a7d07808f924a
Signed-off-by: Dave Schmidt <d.schmidt@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/341538
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Elena Satraitis <w04888@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>