IKQCOMSD1-1295 - Fix calculation of coulomb counter at end of charge
At the end of a charge which results in a full battery, the
coulomb counter state is saved.  At the end of any charge, the BMS
parameter 'bms_end_cc_uah' is supposed to be set to coulomb counter
state.It is calculated by subtracting the CC state at 100% - current
CC state.

However, this calculation is done after saving CC state at 100%.  This
means the value is always 0, since the saved state and current state
are the same.

To avoid this, calculate 'bms_end_cc_uah' before saving the state.

Change-Id: Ib6164e06ad1284309c30bd3fde9cf8f3f4638696
Reviewed-on: http://gerrit.pcs.mot.com/342954
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Kenneth Kessler <w65789@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>