IKHSS7-6172: Adjust MPP12 only on P1 Qinara
The fixes eprom reading on Vanquish with real batteries
Originally the change was made to fix P1 Qinara and it
was verified by HW to not break any other designs.  That
was tested and passed but Vanquish with Real Battery was
not and fails.  This change checks Qinara and P1 to adjust
MPP 12 settings to the change from IKQCOMSD1-1119

Change-Id: Ifc8b7fc99895ccf8fd42ffe51c4c380db9750943
Signed-off-by: w65789 <w65789@motorola.com>
Change-Id: I5d14c2dcd7f8143cabbfde93c635cbb71c341ed0
Reviewed-on: http://gerrit.pcs.mot.com/343942
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>