IKKRNBSP-921 msm8960: mmi: split engineering and production settings
With this change arch/arm/configs/ext_config/eng_bld.config will be
appended to the defconfig file for Android eng and userdebug build
variants. The idea being that additional debugging options can be
enabled for these builds versus what is available in a "production"
build (a "user" build in Android parlance).

Change-Id: I29d59e4afcfed46c2689a758750dd2e72e81d996
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/347296
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>