IKHSS7-14571 resubmit IKHSS7-6741 mmc: msm_sdcc: fix panic during pio transfer
The change was overwritten by similar commit in QCOM 1.0.40
upmerge and then reverted by another commit in QCOM 1.0.40.50
upmerge. Need to re-submit the fix.

The segment in scatter-gather list can be larger than one page.
Currently the mapping is only created for the first page of a
segment. Accessing beyond the page will generate kernel panic.
This change uses the sg iterator to iterate over sg entries so
that the mapping is created page-by-page.
Signed-off-by: Yuanyuan Zhong <a6510c@motorola.com>

Change-Id: I499303d0b658dcd2c4ea4bbb7f2294ad6220ac6d
Reviewed-on: http://gerrit.pcs.mot.com/348288
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Russell Knize <russ.knize@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/373143