IKHSS7-6741 mmc: msm_sdcc: fix panic during pio transfer
The segment in scatter-gather list can be larger than one page.
Currently the mapping is only created for the first page of a
segment. Accessing beyond the page will generate kernel panic.
This change uses the sg iterator to iterate over sg entries so
that the mapping is created page-by-page.
Signed-off-by: Yuanyuan Zhong <a6510c@motorola.com>

Change-Id: I239039d8a7f4f841b5df261da2e37b5760bc4839
Reviewed-on: http://gerrit.pcs.mot.com/348288
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Russell Knize <russ.knize@motorola.com>