IKHSS7-14115 Pull IKHSS6-6825 Dump kernel stack when panic happens.
For the Softlockup "generic_exec_single()" issue, there are
some cases: the CPU is stuck with interrupt disabled. So far,we
haven't good solution to figure out what code was running at that
time on the stuck CPU. This patch is to enable the apanic_mmc to be
able to dump the kernel stack on the stuck CPU when this issue
happens.

This patch should be reverted when this issue is resolved or new
PTM/ETB solution is implemented.

Change-Id: I6dd321297a33809688ffb9b69b3e573a4864fba3
Signed-off-by: Yi-wei Zhao <gbjc64@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/348810
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Tao Hu <taohu@motorola.com>
Reviewed-by: Dong-Dong Yang <gdb746@motorola.com>
Reviewed-by: David Ding <dding@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/371260
Reviewed-by: Yuanyuan Zhong <a6510c@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>