IKHSS7-7008: Detect Accy at Powerup
The Charging driver pm8921-charger.c was not
receiving Accy notifications from a Inserted
Accy at powerup.  This was because the driver
had not gotten to the point of registering with
emu_det accy notifier before emu_det notified
the list.  This commit adds a function in emu_det
to get the current accy so drivers can query
the present accy and defeat the race condition.

Change-Id: I43e3e8774ab5464eea3e5aec13c26da1b76a2484
Signed-off-by: w65789 <w65789@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/350904
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Zachary Crosby <zac@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>