IKHSS7-8271 Fixes Needed For MOT HDMI HAL
In order for the MOT HDMI HAL and stack to be included, two
changes are needed.  The HDMI connection state needs to be
reported in the Android expected way, via a switch state
change.  The raw EDID reading SYSFS needs to be faster, thus
the initially read EDID data is cached and the SYSFS read
returns data from the cache.

Change-Id: Idd72e24d80a5426c13e24f6bb1f0d5dbaf48f12e
Signed-off-by: Steven Fischer <Steven.Fischer@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/351703
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>