IKMMINTG-554 Fix HDMI HPD Switch UEvent Handling
This change will update the HDMI driver to support the
HDMI HPD UEvent in a way that is consistent with support
for non-HDCP devices.  The changes include:

a) Disconnect HDCP status from the HPD uevent, since we
want to support HPD for non-HDCP devices.

b) Allow user space to control when HDCP is enabled.  This
was needed since lockups where occurring if HDCP was in the
process of setting up when user space enabled the HDMI
output.  Previously the HPD uevent did not occur until HDCP
setup was complete.

Change-Id: Idc01e72ad71f8757a522b5db2f1881db51b70eec
Signed-off-by: Steven Fischer <Steven.Fischer@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/357210
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Sumit Bhojesing Jamadar <sumit@motorola.com>
Reviewed-by: Hareesh Gundu <MRFN74@motorola.com>
Reviewed-by: Russell Knize <russ.knize@motorola.com>