IKHSS7-9854 : Release msm_otg wakelock after PPD disconnect
When a PPD is connected the EMU detection driver will trigger a resume of the
MSM OTG driver. Upon disconnection of the PPD a notification is sent to the MSM
OTG driver that nothing is connected. The MSM OTG driver was returning right
away in this case but this causes msm_otg wakelock to be held indefinitely.
Since msm_otg_runtime_resume() increments the device usage count
msm_otg_accy_notify() needs to decrement the usage count before returning in
the IDLE -> IDLE case.

Change-Id: Ibe0b3436050db0305690b88dafe619f2a9f70469
Signed-off-by: Neil Patel <NPATEL@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/357225
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>