IKHSS7-10393: reimplement camera regulator overrides
The msm_camera framework assumes use of several regulators for
camera.  However, this model does not always apply, so we override
(disable) them.

This was done on the initial camera commit, but the code got
lost on one of the recent merges, hence inappropriate regulators
were turning on along with camera.

Change-Id: I1c93706220122d7d4400dc83c9e825dc99d7de60
Signed-off-by: Dave Schmidt <d.schmidt@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/358836
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Elena Satraitis <w04888@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jason Hrycay <jason.hrycay@motorola.com>
Reviewed-by: Russell Knize <russ.knize@motorola.com>