IKHSS7-10968: Add Notifier to notfiy the Audio cable Status to Audio driver
As a part of supporting the EMU audio on docks, the audio driver needs
to know the Audio cable status. This change adds a notifier to inform
the Audio driver the chnage in Audio cable status.

Based on the presence of Audio cable in the dock, the Audio is routed
through the EMU(Dock) or Phone Speaker. The Audio driver would control
the MUX switch to Audio mode OR the default mode, through the EMU detection
driver, depending upon the presence of Audio cable in the dock.

Change-Id: I03206d5239ebc31c50c381a198912eff06e2600c
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/363943
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>