IKVANQUISH-524: fix timing issue between ESD & panel shutdown
There is a issue of the QCOM to read the data from panel, and because
this issue, when the ESD happens, the read will return invalid data.
To get around this problem, we need to read panel data 2 times and because
of the video mode of Vanquish, we have to release the "ov" muxtex for
100msec. At this moment, if the panel can shutdown, and when we
are done with 100msec sleeping, we issue the next read, and we will
have problem to read because the panel was down. This will trigger a
unwanted ESD recovery.

Change-Id: Id4f20f4d173262d1f2e14bb3d5f7a0c5116cb0cc
Signed-off-by: Bang Nguyen <bnguyen1@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/360791
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Limei Wang <e12499@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>