IKHSS7-10862 - Smoke: Cannot detect external sdcard for QNR and Vanquish
During upmerge, part of commit 9090d2 is missed in msm_sdcc.c.
This commit adds it back.

Another change is to disable deferred resume feature. QC driver
doesn't support deffered resume well. If it is enabled, there
will be a lot of warning, and rq timeout. It should be reverted.

Change-Id: I82c286f06469ca0ca43fbf2115f49d7cffc6ff3e
Signed-off-by: Junxiao Chang <jxchang@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/361581
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Dong Zhao <kevin.d.zhao@motorola.com>
Reviewed-by: Xiao-Qin Zhang <QGB647@motorola.com>
Reviewed-by: Valeriy Marchevsky <v.marchevsky@motorola.com>
Reviewed-by: Christopher Fries <C.Fries@motorola.com>
Reviewed-by: Dmitry Ornatsky <cdo021c@motorola.com>
Reviewed-by: Russell Knize <russ.knize@motorola.com>