IKHSS7-10987: Support for HW changes where UART from Krait is routed over GSBI4.
Currently I2C from DSPS and UART from Krait are routed over GSBI12 and
both working concurrently. On the final version of shipping HW, we have
GSBI4 unused. As a result, the UART from Krait could use the GSBI4.
This is to support the HW changes done for this.

This change is backward compatible, where on the older HW, I2C from DSPS
and UART from Krait are routed over GSBI12 is still supported.

Change-Id: I44ed9cd84cc4973c3995ce63ec001e5d2bb51701
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/366891
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>
Reviewed-by: Gregory Meiste <w30289@motorola.com>