IKHSS7-11638 - non-blinking charging LED
Factory needs charging LED to be solid when on.  I have implemented a
blink_set function which is called by led_trigger_blink.  I have also changed
the default trigger to charging-blink-solid-full.  This trigger is set to
blinking when the battery is charging and to solid when it's charged.
Writing the values directly to sysfs file will set it to solid.

Also fixed an issue of mutex_lock warning.  Also renamed the LED to 'charging'
to simplify testing.

Change-Id: I2ef5b3211e433af8a30016b31c98c1ce59a6025c
Signed-off-by: Alina Yakovleva <qvdh43@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/365943
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Igor Kovalenko <Igor.Kovalenko@motorola.com>