IKHSS7-11676 : Disable USB phy reset on disconnect
Resetting the USB phy on disconnect was a workaround QC used on previous
chips for issues that are no longer present on the MSM8960 according to
the QC HW team.

There have been cases where the reset leads to ULPI write timeouts if the
controller is already in low power mode. Since the reset isn't required
and slows down disconnect processing it will be disabled.

Change-Id: I31e3fa4fb79244dee9bca4333f0aea77d988e361
Signed-off-by: Neil Patel <NPATEL@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/367910
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Konstantin Makariev <hcv867@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>