IKMMINTG-613 Fix HDMI Black Frame
This change fixes an issue where a timer expires even though
it was not directly needed, causing the HDMI output to shut
off for a moment.  The timer is used somewhat strangely, in
that it is added at probe timer and assumed to stay on the
timer queue and only mod'd when needed.  Since the timer is
sitting on the queue, it can expire and does, even thought it
was not needed, what results in an unexpected behavior.  This
change adds a flag to allow the timer function to filter the
response to only those expiration that were desired.

Change-Id: I2bf32b1a048efa49ed9421cbd901bd198494643b
Signed-off-by: Steven Fischer <Steven.Fischer@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/369961
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>