IKHSS7-6601: Add Alarm to Wakeup Battery Driver
The AP can sleep while the BP is still in an active call. The Temperature and
State of Charge can not be tracked from Sleep so we must wakeup periodically
and update User Space.  This will allow both Cooldown mode and Low Battery
warnings to occur so the user can take action.

Change-Id: Ic8d43d7c056e6321d12b4c5d819ae54aa160854e
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Change-Id: Ib86f954eb34ac1ffecd51d41ba4b484b618bb678
Reviewed-on: http://gerrit.pcs.mot.com/369971
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Kenneth Kessler <w65789@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Jason Knopsnyder <jasonk@motorola.com>