IKHSS7-14849: Enabling L2 cache error reporting breaking UART communication.
By enabling L2 cache error reporting flags, we see the L2 master port slave
error happening when we attempt to do UART communication. This is in turn
breaking Whisper(UART) authentication.

A SR has been opened, and have asked Qualcomm to provide a fix for this.
Until then this flag would be disabled.

Change-Id: If3593138faee0e92ac80dd6d81b6fefba8de7e71
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/372141
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>