IKHSS7-16016: Fix ethernet dongle disconnect hang
When disconnecting the ASIX USB Ethernet dongle the unregister()
was blocked waiting for flush_scheduled_work() to complete. The
function was being called from a worker context which made this
unsafe.

Instead of using flush_scheduled_work() we can cancel the
specific work item that needs to be flushed. This change has
been applied to the other usbnet drivers via this commit:

   23f333a drivers/net: don't use flush_scheduled_work()

Change-Id: I0a474c33fee851f9e2ebac69e548458058679c38
Signed-off-by: Neil Patel <NPATEL@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/374761
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>