IKMMINTG-676 Handle HDMI BG Audio Use Case
This change fixes a bug which allowed the HDMI output to freeze
on the last outputted frame when entering the HDMI background
audio use case.  The bug is that the overlay code did not
disable the HDMI mixer upon the UNSET command, instead is waited
for one more frame to be output before the mixer was disabled.
This caused the last frame to freeze on the display since in the
background audio case there were no further UI frames to be sent.

Change-Id: I6dba56f5e182dfd8c6eeda283ba603621ab43204
Signed-off-by: Steven Fischer <Steven.Fischer@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/377706
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>