IKHSS7-16616 Support for mapping HOB RAM Addresses
Provide mmap support on HOB shared memory.
The virtual memory will be marked as non-cached memory.

HOB or the HandOver Block is the staging area for Phasing Data.
Since modem does not have direct access to eMMC, the HOB data
is copied to HOB shared RAM area and this data will be written
to the HOB partition by the apps processor.

Change-Id: I200cc02ecfb60f63df3d704fe2cdb5546787fc14
Signed-off-by: Satish D <SDoraiswamy@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/379189
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>