IKHSS7-18902: Reset and Clear the Safety timer when the userspace is updated.
The charger was observed to freeze for unknown reasons.
So the QCOM recommended us to clear the charge failed bit upon
insertion/removal of usb, in eoc check worker and the heartbeat worker.

In testing I saw the charge fail interrupt getting triggered when the
safety timer expires where we clear the charge failed bit. As a result,
the above workaround is no longer needed. We also reset and clear the
timer in update_heartbeat i.e. internal function to update userspace.

Change-Id: I0415d88d82d47eb58dc529f58b3ed8684a4993e9
Signed-off-by: Vidyadhar Gurram <rnh786@motorola.com>
Reviewed-on: http://gerrit.pcs.mot.com/381821
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Keyur Khambholja <Keyur.Khambholja@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>