IKHSS7-19058: Vanquish video mode image fade away when suspend phone.
The issue happens on video mode panel because QCOM video_off shutdown
MDP block and then dsi_off->panel_off, because data tx is stopped
when display is still on, which caused the fading.

To minimize the change, send DISP off cmd to disable panel output, so
the sequence happens in darkness.

Change-Id: I9bf29f5880c4cad8ca37bb9d0b1ac69d2777cfc7
Signed-off-by: Limei Wang <e12499@motorola.com>
Change-Id: I53e3d955cc48be674b7f61033b5c1bd9bcc282e5
Reviewed-on: http://gerrit.pcs.mot.com/382225
Tested-by: Jira Key <JIRAKEY@motorola.com>
Reviewed-by: Check Patch <CHEKPACH@motorola.com>
Reviewed-by: Klocwork kwcheck <klocwork-kwcheck@sourceforge.mot.com>
Reviewed-by: Bang Nguyen <bang.nguyen@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>